% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanImmuneExpViewer.R
\name{ExpressionViewer}
\alias{ExpressionViewer}
\title{Create a shiny app to display expression values and/or alternative splicing}
\usage{
ExpressionViewer(
  experiment = counts,
  metaNames = "label.main",
  printNames = " ",
  paper = " ",
  paperlink = " ",
  apptitle = "Expression Viewer:"
)
}
\value{
A shiny app displaying the expression (and potential splicing) of
the experiment input
}
\description{
This function takes a SummarizedExperiment object, SGFeatureCounts object, or
a list of either of these classes and produces a shiny app with a barplot
illustrating expression levels per gene and (in the case of SGFeatureCounts)
a heatmap illustrating alternative splicing per sample. The assumptions for
each object case are listed below.
}
\details{
Summarized Experiment:
The sample information wishing to be displayed as the independent variable
is in the colData with the column name inputted into the variable "metaNames".
The rownames of the expression matrix are gene names.

SGFeatureCounts Object:
The sample information wishing to be displayed as the independent variable
is in the colData with the column name inputted into the variable "metaNames".
The geneID() of the experiment is complete with the gene annotation style
desired (NCBI, HGNC Symbol, etc.).

List of either SummarizedExperiments or SGFEatureCounts:
The "metaNames" are consistently named for each element of the list. Any
class assumptions are retained when part of a list.
}
